requirements.txt
flask
psycopg2-binary
cloudinary
gunicorn
werkzeug
